+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Changed
 - fix missing README.rst
 - handle input counts arrays > 1D (always flatten to 1D)
+- handle single-class case
 
 ## [0.5.3] - 2018-03-15
 ### Changed